requirements.txt
pyirsdk
pandas
numpy
PyQt6
pyqtgraph
